third_party/WebKit/LayoutTests/fast/dom/Window/getMatchedCSSRules-nested-rules-expected.txt

CONSOLE WARNING: line 1: document.getMatchedCSSRules() is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4606972603138048 for more details.
getMatchedCSSRules with nested CSS rules

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getMatchedCSSRules(document.querySelector('div'))[0].cssText is "div { color: blue; }"
PASS getMatchedCSSRules(document.querySelector('.test-element'))[0].cssText is ".test-element { border: 2px solid black; }"
PASS successfullyParsed is true

TEST COMPLETE
Test getMatchedCSSRules to output deeply nested CSSRules.

All this text should be blue.
This span should have border.
